Add books, Retrieve books and view users
The add books form adds the new book data in to the book table in the
database. Also on the browse books page for regular users, the
controller retrieves data from the books table and displays it. The
admin can now see all the registered users on the website on the manage
users page.