Used if statements to hide and show different buttons and views
I had 2 layouts before which caused too many problems. Now I used if
statements to show that if the user is admin then show admin views and
buttons on the dashboard. But if the user is not admin then show items,
buttons and views that are for the normal user.