Edit button works partially
The edit button on the manage books page redirects the use to the edit
page where the form retrieves the data of the row and places it in the
text boxes. But css and update button do not work for some reason. Also
jqueries are running in the background when the edit form loads up, this
slows the whole page down.