Added hyperlink to book titles redirecting to new view
Implemented a new view which will redirect the user once the book title
hyperlink is clicked on the "Browse Books" view. In that view,
information on the chosen book will be presented. Also the standard user
dash board contains interactive image buttons that will redirect them to
certain views similar to the admin dashboard.