Comment Form Work Completely
The comment form on the bookinfo view now submits the id of the user, id
of the book and id of the comment (along with the comment) on to the
database. Now to get the comments to view on the book where they were
commented on; a 3 way inner join needs to be implemented on the models.