Implemented Pagination and Show More/Less Text
In browse books and Managebooks view, the items that are to appear on
the pages are now limited. This means that the rest of the items that
were to appear will now appear in other pages due to the pagination
method. Also on the bookinfo and managebooks view, the book descriptions
now have show less and more feature to increase or reduce text.