Implementing the voting feature
Created a votes table which will contain the likes and dislikes made by
users regarding a book. There are columns like user_id and book_id which
will be used as foriegn keys to link the users and books table. So far
the votes by users are saved on the database such as the user_id and the
book_id as well as the voted choice (like, dislike). On the bookinfo
view, all the votes are shown but they are shown as 1 and 0's in a
foreach loop.