Voting Works (Has few bugs)
The voting system now works. On the book view page the user can see the
number of up and down votes on a view. But on the developers side, when
the up or download icon is clicked on the browsebooks view, the wrong
book_id is submitted in to the database.