Implemented Password Reset Feature
If a user forgets their password for the website, they can now use the
password reset link on the login page. If the email address in the
password reset form is not found in the database then it will show a
message but if the email address is matched with the one in the database
then the link will be sent to that email address. Once the user receive
the password reset link and clicks it,  they will redirected to the
bookrepo website where they can insert the new password link.