Implemented Role Middleware and Custom Error View
The role middleware will make sure only the admin can access the
administrator features like manage books, users and comments etc. It
also prevents the standard user and guest user from accessing those
pages via the URL search bar.

In addition, a custom error page is made to be displayed if the user
uses the URL search bar to search a page that does not exist on the
website.