Manage Users View Improvements
In the manage users controller, the "getusers" method now displays all
the users who are not admin. If the admin was there and was accidently
banned then that could cause problems. An if statements is also
implemented on the manage users view which displays the ban button if
the user is not banned and displays the unbanned button when the user is
banned.