Implemented Ban Middleware
Created a Ban Middle which restricts the banned user from accessing
views other than the contact view. If the banned user tries to put the
URL of the views on the search bar then the page will be redirected back
to the dashboard containing the Ban notification. In addition to this,
the Auth middle is also used to stop guest users form using  registered
users features such as request books view and discussions view.