Added a graph on statistics view and more changes
On the statistics view, the total count of data in the tables are now
displayed in a graph. Also on the add books and edit books view, the
admin can select a genre from a dropdown rather than typing it in. In
addition to these changes, more books have been added to the database
which can be seen in the browse books view.