Implemented A Genre Search Form
On the browse books view, the user can now search for books based on the
genre. Once submitted, it will redirect the user to the genre search
results view where they can see books based on the genre. The title of
those are hyperlinked which will redirect them to the book of the users
choice.