New View and Fixed Topic Delete Method
The topic delete button on the discussions view was not working due to
having two delete methods on the same route. So now the creator of the
topic can delete their topics by going to the manage topics view in the
Forums dropdown link. The delete from the discussions view is no longer
needed.